+++ src/utils.cpp
@@ -1,7 +1,10 @@
 #include <memory>
 #include <string>
+#include <future>
 #include "utils.h"
-
+#if !defined(_WIN32) && !defined(_WIN64)
+#include <signal.h>
+#endif
 namespace odbc {
 namespace utils {
 
@@ -47,4 +50,79 @@ namespace utils {
       }
     }
   }
+
+  void run_interruptible(const std::function<void()>& exec_fn, const std::function<void()>& cleanup_fn)
+  {
+    std::exception_ptr eptr;
+#if !defined(_WIN32) && !defined(_WIN64)
+    sigset_t set, old_set;
+    sigemptyset(&set);
+    sigaddset(&set, SIGINT);
+    int rc = pthread_sigmask(SIG_BLOCK, &set, &old_set);
+    if ( rc != 0 )
+    {
+      // Unable to properly mask SIGINT from execution thread
+      raise_warning("Unexpected behavior when creating execution thread.  Signals to interrupt execution may not be caught.");
+    }
+#endif
+    auto future = std::async(std::launch::async, [&exec_fn, &eptr]() {
+      try {
+        exec_fn();
+      } catch (...) {
+        eptr = std::current_exception();
+      }
+      return;
+    });
+#if !defined(_WIN32) && !defined(_WIN64)
+    pthread_sigmask(SIG_SETMASK, &old_set, NULL);
+#endif
+    std::future_status status;
+    do {
+      status = future.wait_for(std::chrono::seconds(1));
+      if (status != std::future_status::ready) {
+        try { Rcpp::checkUserInterrupt(); }
+        catch (const Rcpp::internal::InterruptedException& e) {
+          raise_message("Caught user interrupt, attempting a clean exit...");
+          cleanup_fn();
+        } catch (...) { throw; }
+      }
+    } while (status != std::future_status::ready);
+    if (eptr) {
+      // An exception was thrown in the thread
+      try { std::rethrow_exception(eptr); }
+      catch (const odbc_error& e) { raise_error(e); }
+      catch (...) { raise_message("Unknown exception while executing"); throw; };
+    }
+  }
+
+  void raise_message(const std::string& message) {
+    Rcpp::Environment pkg = Rcpp::Environment::namespace_env("cli");
+    Rcpp::Function r_method = pkg["cli_inform"];
+    Rcpp::CharacterVector argMessage =
+      Rcpp::CharacterVector::create(Rcpp::Named("i", message));
+    r_method(argMessage);
+  }
+
+  void raise_warning(const std::string& message) {
+    Rcpp::Environment pkg = Rcpp::Environment::namespace_env("cli");
+    Rcpp::Function r_method = pkg["cli_warn"];
+    Rcpp::CharacterVector argMessage =
+      Rcpp::CharacterVector::create(Rcpp::Named("!", message));
+    r_method(argMessage);
+  }
+
+  void raise_error(const std::string& message) {
+    Rcpp::Environment pkg = Rcpp::Environment::namespace_env("cli");
+    Rcpp::Function r_method = pkg["cli_abort"];
+    Rcpp::CharacterVector argMessage =
+      Rcpp::CharacterVector::create(Rcpp::Named("x", message));
+    r_method(argMessage);
+  }
+
+  void raise_error(const odbc_error& e) {
+    Rcpp::Environment pkg = Rcpp::Environment::namespace_env("odbc");
+    Rcpp::Function r_method = pkg["rethrow_database_error"];
+    r_method(e.what());
+  }
+
 }}